+++ src/main/java/Databases/RestaurantDatabase.csv
@@ -1,4 +1,4 @@
 location,name,cuisineType,priceBucket,owner,avgStars,reviews
 K9B 5L0,Ajisen,Ramen,3,222,0.0,empty
 M1B 1M2,Tim Hortons,Coffee,1,111,0.0,empty
-Z1B 2C8,Chako,K BBQ,5,333,0.0,empty
+Z1B 2C8,Chako,Food,5,333,0.0,empty
