--- src/main/java/filtering_use_case/ChoicesDSGateway.java
@@ -1,9 +1,0 @@
-package filtering_use_case;
-
-import entities.Restaurant;
-
-import java.util.ArrayList;
-
-public interface ChoicesDSGateway {
-    ArrayList<Restaurant> retrieveAllRestaurants();
-}
