--- src/main/java/filtering_use_case/ChoicesInputBoundary.java
@@ -1,5 +1,0 @@
-package filtering_use_case;
-
-public interface ChoicesInputBoundary {
-    ChoicesResponseModel select(ChoicesRequestModel requestModel);
-}
