--- src/main/java/filtering_use_case/ChoicesInteractor.java
@@ -1,8 +1,0 @@
-package filtering_use_case;
-
-public class ChoicesInteractor {
-
-    // File path
-    // String restaurants_file = "https://github.com/CSC207-2022F-UofT/course-project-group-72/blob/76fc88b7cc4029" +
-    //            "ba7b25b3d9fec42059602721f8/src/main/java/databases/restaurants.csv";
-}
