--- src/main/java/filtering_use_case/ChoicesResponseFormatter.java
@@ -1,15 +1,0 @@
-package filtering_use_case;
-
-public class ChoicesResponseFormatter implements ChoicesPresenter{
-
-    @Override
-    public ChoicesResponseModel prepareSuccessView(ChoicesResponseModel responseModel){
-        // TODO: success screen
-
-        return responseModel;
-
-    }
-
-    // TODO: prepareFailView
-
-}
