+++ src/main/java/filtering_use_case/screens/ChoicesPresenter.java
@@ -1,8 +1,8 @@
-package filtering_use_case;
-
-import filtering_use_case.ChoicesResponseModel;
+package filtering_use_case.screens;
 
 public interface ChoicesPresenter {
     ChoicesResponseModel prepareSuccessView(ChoicesResponseModel responseModel);
+
+    ChoicesResponseModel prepareFailView(String error);
 
 }
