+++ src/main/java/filtering_use_case/screens/ChoicesSortedView.java
@@ -1,4 +1,4 @@
-package filtering_use_case;
+package filtering_use_case.screens;
 import entities.Restaurant;
 import entities.RestaurantFactory;
 import entities.User;
